Add ssh backend
It is a first proposal to discuss if I'm on the right track.
If the SUT is only accessible via SSH, then this backend comes
in place (e.g. public cloud instances).

Implementation:
Use three ssh connections to SUT.
The first connection is used to setup the SUT. Create
a named pipe, which will be later used to read the return
code from commands.
The second connection is used to read from that named pipe
using 'tail -f'.
The third connection launch the shell where commands get executed.
The virtio_screen.pm implementation is currently used to do that.

Current limitations:
 * Only text console, no GUI
 * No serial output from linux
 * No ssh key authentication
 * Executed scripts are not logged.
 * Not tested with other consoles like localXvnc
 * Use sudo to gain root