distribution: Use wait_screen_change in script_output
We encounter a stability problem with simple `script_output('w')` call.
1 out of ~50 calls failed, because the content of the script is
```
w\r
```
instead of
```
w\n
```

The actual root cause is unknown. But it seems that we encounter similar
problems as fixed with 87f895f3, so we use `wait_screen_change()` to
get some change.
The behavior was only seen on hyperv and vmware hosts.

Ticked: https://progress.opensuse.org/issues/60566